requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-environ==0.8.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==11.1.0
docxtpl==0.15.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
inline==0.0.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
null==0.6.1
packaging==21.0
Pillow==8.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.3.0
pyparsing==3.0.0
pytz==2021.3
PyYAML==6.0
pytest==6.2.5
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
self==2020.12.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
